requirements.txt
tensorflow-cpu==2.10.0
flask
gunicorn
numpy==1.23.5
pillow
pandas
joblib

